LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/memory64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:4"
FAIL #3 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:5"
FAIL #4 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:6"
FAIL #5 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:7"
FAIL #6 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:8"
FAIL #7 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:9"
FAIL #8 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:11"
FAIL #9 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:13"
FAIL #10 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:15"
PASS #11 Test that WebAssembly compilation fails
PASS #12 Test that WebAssembly compilation fails
PASS #13 Test that WebAssembly compilation fails
PASS #14 Test that WebAssembly compilation fails
PASS #15 Test that WebAssembly compilation fails
PASS #16 Test that WebAssembly compilation fails
PASS #17 Test that WebAssembly compilation fails
PASS #18 Test that WebAssembly compilation fails
PASS #19 Test that WebAssembly compilation fails
PASS #20 Test that WebAssembly compilation fails
PASS #21 Test that WebAssembly compilation fails
PASS #22 Test that WebAssembly compilation fails
PASS #23 Test that WebAssembly compilation fails
PASS #24 Test that WebAssembly compilation fails
FAIL #25 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "memory64.wast:71"
FAIL #26 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) true but got (string) "wrapper"
PASS #27 Reinitialize the default imports
FAIL #28 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 16: Memory64 is not enabled at {loc} expected true got false
FAIL #29 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #30 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 16: Memory64 is not enabled at {loc} expected true got false
FAIL #31 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:13:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #32 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 16: Memory64 is not enabled at {loc} expected true got false
FAIL #33 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:19:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #34 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 16: Memory64 is not enabled at {loc} expected true got false
FAIL #35 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:25:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #36 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 16: Memory64 is not enabled at {loc} expected true got false
FAIL #37 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 16: Memory64 is not enabled at {loc} expected true got false
FAIL #38 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:34:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #39 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 35: Memory64 is not enabled at {loc} expected true got false
FAIL #40 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:40:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #41 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:43:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #42 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 35: Memory64 is not enabled at {loc} expected true got false
FAIL #43 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:49:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #44 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:52:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #45 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 35: Memory64 is not enabled at {loc} expected true got false
FAIL #46 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:58:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #47 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
PASS #48 Test that WebAssembly compilation fails
PASS #49 Test that WebAssembly compilation fails
PASS #50 Test that WebAssembly compilation fails
PASS #51 Test that WebAssembly compilation fails
PASS #52 Test that WebAssembly compilation fails
PASS #53 Test that WebAssembly compilation fails
PASS #54 Test that WebAssembly compilation fails
PASS #55 Test that WebAssembly compilation fails
PASS #56 Test that WebAssembly compilation fails
PASS #57 Test that WebAssembly compilation fails
PASS #58 Test that WebAssembly compilation fails
PASS #59 Test that WebAssembly compilation fails
PASS #60 Test that WebAssembly compilation fails
PASS #61 Test that WebAssembly compilation fails
FAIL #62 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 60: Memory64 is not enabled at {loc} expected true got false
FAIL #63 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:109:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #64 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:112:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #65 Test that WebAssembly compilation succeeds assert_true: module@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:146:24
@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:115:31
run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:224:37
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:115:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got "wrapper"
FAIL #66 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:115:24
run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:224:37
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:115:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #67 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:115:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #68 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:118:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #69 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:121:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #70 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:124:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #71 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:127:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #72 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:130:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #73 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:133:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #74 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:136:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #75 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:139:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #76 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:142:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #77 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:145:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #78 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:148:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #79 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:151:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #80 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:154:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #81 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:157:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #82 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:160:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #83 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:163:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #84 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:166:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #85 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:169:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #86 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:172:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #87 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:175:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #88 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:178:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #89 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:181:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #90 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:184:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #91 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:187:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #92 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:190:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #93 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:193:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #94 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:196:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #95 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:199:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #96 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:202:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #97 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:205:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #98 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:208:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #99 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:211:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #100 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:214:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #101 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:217:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #102 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:220:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #103 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:223:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #104 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:226:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #105 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:229:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #106 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:232:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
FAIL #107 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
memory64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:235:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
